+++ vehicle/src/Vehicle/Backend/ITP/Rocq.hs
@@ -45,7 +45,8 @@ import Vehicle.Data.Variable.Bound.Context.Name
 
 data RocqOptions = RocqOptions
   { output :: Maybe FilePath,
-    moduleName :: Maybe String
+    moduleName :: Maybe String,
+    constructiveReals :: Bool
   }
 
 currentPhase :: Doc ()
@@ -59,12 +60,19 @@ compileProgToRocq prog _options =
     -- Collects dependencies by first discarding precedence info and then
     -- folding using Set Monoid
     let programDependencies = fold (reAnnotateS fst programStream)
+    -- If using constructive reals, add the required import
+    let programDependencies' =
+          if constructiveReals _options
+            then
+              Set.insert (MathcompImport Rstruct) $
+                Set.insert (RequireImport ConstructiveReals) programDependencies
+            else programDependencies
 
     let rocqProgram =
           unAnnotate
             ( (vsep2 :: [Code] -> Code)
-                [ importStatements programDependencies,
-                  preamble programDependencies,
+                [ importStatements programDependencies',
+                  preamble programDependencies',
                   programDoc
                 ]
             )
@@ -122,6 +130,7 @@ data Mathcomp
   = Boot
   | Algebra
   | Reals
+  | Rstruct
   deriving (Eq, Ord)
 
 instance Pretty Mathcomp where
@@ -129,16 +138,19 @@ instance Pretty Mathcomp where
     Boot -> "all_boot"
     Algebra -> "all_algebra"
     Reals -> "all_reals"
+    Rstruct -> "Rstruct"
 
 data Library
   = VehicleTensor
   | VehicleUtils
+  | ConstructiveReals
   deriving (Eq, Ord)
 
 instance Pretty Library where
   pretty = \case
     VehicleTensor -> "vehicle.tensor"
     VehicleUtils -> "vehicle.utils"
+    ConstructiveReals -> "Stdlib.Reals.Reals"
 
 data RocqModule
   = OrderDef
@@ -164,10 +176,12 @@ importStatements :: Set Dependency -> Code
 importStatements deps = vsep $ map pretty (Set.toList deps)
 
 preamble :: Set Dependency -> Code
-preamble deps =
-  if Set.member (MathcompImport Reals) deps
-    then "Parameter" <+> "R" <+> ":" <+> align "realType" <> "."
-    else ""
+preamble deps
+  | Set.member (RequireImport ConstructiveReals) deps =
+      "Notation" <+> "R" <+> ":=" <+> align "Rdefinitions.R" <> "."
+  | Set.member (MathcompImport Reals) deps =
+      "Parameter" <+> "R" <+> ":" <+> align "realType" <> "."
+  | otherwise = ""
 
 --------------------------------------------------------------------------------
 -- Intermediate results of compilation
@@ -514,7 +528,7 @@ compileBuiltin b args = case b of
     QuantifyRatTensor q -> case reverse args of
       (ExplicitArg _ (Lam _ binder body)) : _ -> compileTypeLevelQuantifier q [binder] body
       _ -> unsupportedArgsError
-    AtTensor -> compileNotationAndArgs [RequireImport VehicleTensor] LeftAssociative (Just 2) "$0 ^^ $1" (Just "nindex") args
+    AtTensor -> compileNotationAndArgs [RequireImport VehicleTensor] LeftAssociative (Just (-2)) "$0 ^^ $1" (Just "nindex") args
     If -> compileNotationAndArgs [MathcompImport Boot] NotAssociative (Just 0) "if $0 then $1 else $2" Nothing args
     ForeachTensor -> compileApplication [RequireImport VehicleTensor] "nstack" args
     StackTensor -> compileStack args
@@ -627,11 +641,14 @@ bracketArgs maybeParentPrecedence = traverse bracketArg
       logDebug MaxDetail $
         "!!!"
           <+> body
-          <+> pretty (getPrecedence body)
-          <+> ">="
-          <+> pretty maybeParentPrecedence
-          <+> "="
-          <+> pretty (getPrecedence body >= maybeParentPrecedence)
+          <+> parens
+            ( "precedence"
+                <+> pretty (getPrecedence body)
+                <+> ">="
+                <+> pretty maybeParentPrecedence
+                <+> "="
+                <+> pretty (getPrecedence body >= maybeParentPrecedence)
+            )
       return $ case visibilityOf arg of
         Instance {} -> annotate (mempty, Nothing) $ braces (braces body)
         Implicit {} -> annotate (mempty, Nothing) $ braces body